Moved Dialogs/Modals to new folder
