+++ src/components/dialogs/ParticipantDialog.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { Button, FormGroup } from '@material-ui/core';
-import { IParticipant } from '../data/Model';
+import { IParticipant } from '../../data/Model';
 import Dialog from '@material-ui/core/Dialog';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
